small trace processor refactoring
