+++ cloud/blockstore/libs/daemon/common/bootstrap.cpp
@@ -87,7 +87,8 @@
 #include <cloud/storage/core/libs/diagnostics/logging.h>
 #include <cloud/storage/core/libs/diagnostics/monitoring.h>
 #include <cloud/storage/core/libs/diagnostics/stats_updater.h>
-#include <cloud/storage/core/libs/diagnostics/trace_reader.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor_mon.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor.h>
 #include <cloud/storage/core/libs/diagnostics/trace_serializer.h>
 #include <cloud/storage/core/libs/grpc/logging.h>
 #include <cloud/storage/core/libs/grpc/threadpool.h>
@@ -238,14 +239,15 @@ void TBootstrapBase::Init()
 
     auto diagnosticsConfig = Configs->DiagnosticsConfig;
     if (TraceReaders.size()) {
-        TraceProcessor = CreateTraceProcessor(
-            Timer,
-            BackgroundScheduler,
-            Logging,
+        TraceProcessor = CreateTraceProcessorMon(
             Monitoring,
-            "BLOCKSTORE_TRACE",
-            NLwTraceMonPage::TraceManager(diagnosticsConfig->GetUnsafeLWTrace()),
-            TraceReaders);
+            CreateTraceProcessor(
+                Timer,
+                BackgroundScheduler,
+                Logging,
+                "BLOCKSTORE_TRACE",
+                NLwTraceMonPage::TraceManager(diagnosticsConfig->GetUnsafeLWTrace()),
+                TraceReaders));
 
         STORAGE_INFO("TraceProcessor initialized");
     }
