+++ cloud/blockstore/libs/disk_agent/bootstrap.cpp
@@ -40,7 +40,8 @@
 #include <cloud/storage/core/libs/diagnostics/logging.h>
 #include <cloud/storage/core/libs/diagnostics/monitoring.h>
 #include <cloud/storage/core/libs/diagnostics/stats_updater.h>
-#include <cloud/storage/core/libs/diagnostics/trace_reader.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor_mon.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor.h>
 #include <cloud/storage/core/libs/diagnostics/trace_serializer.h>
 #include <cloud/storage/core/libs/features/features_config.h>
 #include <cloud/storage/core/libs/kikimr/actorsystem.h>
@@ -212,14 +213,15 @@ void TBootstrap::Init()
 
     auto diagnosticsConfig = Configs->DiagnosticsConfig;
     if (TraceReaders.size()) {
-        TraceProcessor = CreateTraceProcessor(
-            Timer,
-            Scheduler,
-            Logging,
+        TraceProcessor = CreateTraceProcessorMon(
             Monitoring,
-            "BLOCKSTORE_TRACE",
-            NLwTraceMonPage::TraceManager(diagnosticsConfig->GetUnsafeLWTrace()),
-            TraceReaders);
+            CreateTraceProcessor(
+                Timer,
+                Scheduler,
+                Logging,
+                "BLOCKSTORE_TRACE",
+                NLwTraceMonPage::TraceManager(diagnosticsConfig->GetUnsafeLWTrace()),
+                TraceReaders));
 
         STORAGE_INFO("TraceProcessor initialized");
     }
