+++ cloud/storage/core/libs/diagnostics/trace_processor_mon.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "public.h"
+
+namespace NCloud {
+
+////////////////////////////////////////////////////////////////////////////////
+
+ITraceProcessorPtr CreateTraceProcessorMon(
+    IMonitoringServicePtr monitoring,
+    ITraceProcessorPtr impl);
+
+}   // namespace NCloud
