+++ cloud/storage/core/libs/diagnostics/trace_reader.h
@@ -2,25 +2,15 @@
 
 #include "public.h"
 
-#include <cloud/storage/core/libs/common/startable.h>
-#include <cloud/storage/core/libs/diagnostics/monitoring.h>
 #include <cloud/storage/core/protos/media.pb.h>
 #include <cloud/storage/core/protos/trace.pb.h>
 
-#include <library/cpp/containers/ring_buffer/ring_buffer.h>
 #include <library/cpp/lwtrace/log.h>
-#include <library/cpp/monlib/service/pages/html_mon_page.h>
-#include <library/cpp/monlib/service/pages/index_mon_page.h>
-#include <library/cpp/monlib/service/pages/templates.h>
-#include <library/cpp/protobuf/util/pb_io.h>
 
-#include <util/datetime/base.h>
-#include <util/generic/vector.h>
-#include <util/system/thread.h>
+#include <util/generic/hash.h>
 
 namespace NLWTrace {
-    class TManager;
-    class TQuery;
+class TManager;
 }   // namespace NLWTrace
 
 namespace NCloud {
@@ -44,7 +34,7 @@ TRequestThresholds ConvertRequestThresholds(
     const TProtoRequestThresholds& value);
 
 void OutRequestThresholds(
-IOutputStream& out,
+    IOutputStream& out,
     const NCloud::TRequestThresholds& value);
 
 struct TEntry
@@ -59,37 +49,17 @@ struct ITraceReader
 {
     const TString Id;
 
-    ITraceReader(TString id)
+    explicit ITraceReader(TString id)
         : Id(std::move(id))
     {}
 
     virtual ~ITraceReader() = default;
 
-    virtual void ForEachTraceLog(std::function<void (const TEntry&)> fn) = 0;
-    virtual void Push(TThread::TId, const NLWTrace::TTrackLog&) = 0;
+    virtual void Push(TThread::TId tid, const NLWTrace::TTrackLog&) = 0;
+    virtual void ForEach(std::function<void(const TEntry&)> fn) = 0;
     virtual void Reset() = 0;
 };
-
 ////////////////////////////////////////////////////////////////////////////////
-
-struct ITraceProcessor
-    : public IStartable
-{
-    virtual ~ITraceProcessor() = default;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
-ITraceProcessorPtr CreateTraceProcessor(
-    ITimerPtr timer,
-    ISchedulerPtr scheduler,
-    ILoggingServicePtr logging,
-    IMonitoringServicePtr monitoring,
-    TString componentName,
-    NLWTrace::TManager& lwManager,
-    TVector<ITraceReaderPtr> readers);
-
-ITraceProcessorPtr CreateTraceProcessorStub();
 
 ITraceReaderPtr CreateTraceLogger(
     TString id,
@@ -110,8 +80,6 @@ NLWTrace::TQuery ProbabilisticQuery(
     const TVector<std::tuple<TString, TString>>& probes,
     ui32 samplingRate,
     ui32 shuttleCount);
-
-bool ReaderIdMatch(const TString& traceType, const TString& readerId);
 
 TDuration GetThresholdByRequestType(
     const NProto::EStorageMediaKind mediaKind,
