+++ cloud/vm/blockstore/bootstrap.cpp
@@ -25,7 +25,8 @@
 #include <cloud/storage/core/libs/diagnostics/logging.h>
 #include <cloud/storage/core/libs/diagnostics/monitoring.h>
 #include <cloud/storage/core/libs/diagnostics/stats_updater.h>
-#include <cloud/storage/core/libs/diagnostics/trace_reader.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor_mon.h>
+#include <cloud/storage/core/libs/diagnostics/trace_processor.h>
 #include <cloud/storage/core/libs/grpc/logging.h>
 #include <cloud/storage/core/libs/grpc/threadpool.h>
 #include <cloud/storage/core/libs/grpc/utils.h>
@@ -139,14 +140,15 @@ void TBootstrap::Init()
     }
 
     if (TraceReaders.size()) {
-        TraceProcessor = CreateTraceProcessor(
-            Timer,
-            Scheduler,
-            Logging,
+        TraceProcessor = CreateTraceProcessorMon(
             Monitoring,
-            "BLOCKSTORE_TRACE",
-            TLWTraceSafeManager::Instance(),
-            TraceReaders);
+            CreateTraceProcessor(
+                Timer,
+                Scheduler,
+                Logging,
+                "BLOCKSTORE_TRACE",
+                TLWTraceSafeManager::Instance(),
+                TraceReaders));
     }
 
     auto rootGroup = Monitoring->GetCounters()
